Split broker abstraction
