+++ pkg/vehicle/vehicle.go
@@ -2,24 +2,7 @@ package vehicle
 
 import (
 	"github.com/HyperloopUPV-H8/h9-backend/pkg/abstraction"
-	"github.com/HyperloopUPV-H8/h9-backend/pkg/transport"
 )
-
-type BrokerTopic string
-type BrokerPush interface {
-	Topic() BrokerTopic
-}
-type BrokerRequest interface {
-	Topic() BrokerTopic
-}
-type BrokerResponse interface {
-	Topic() BrokerTopic
-}
-
-type Broker interface {
-	Push(BrokerPush) error
-	Pull(BrokerRequest) (BrokerResponse, error)
-}
 
 type BoardEvent string
 type BoardNotification interface {
@@ -46,12 +29,8 @@ type Logger interface {
 }
 
 type Vehicle struct {
-	broker    Broker
+	broker    abstraction.Broker
 	boards    map[BoardId]Board
 	transport abstraction.Transport
 	logger    Logger
-}
-
-func (vehicle *Vehicle) sendPacket(packet abstraction.Packet) error {
-	return vehicle.transport.SendMessage(transport.NewPacketMessage(packet))
 }
